--- drive/src/contract/defaults.rs
@@ -1,4 +1,0 @@
-// pub const DEFAULT_CONTRACT_KEEPS_HISTORY: bool = false;
-// pub const DEFAULT_CONTRACT_MUTABILITY: bool = true;
-// pub const DEFAULT_CONTRACT_DOCUMENTS_KEEPS_HISTORY: bool = false;
-// pub const DEFAULT_CONTRACT_DOCUMENT_MUTABILITY: bool = true;
