+++ drive/src/contract/document.rs
@@ -3,32 +3,19 @@ use std::io::{BufReader, Read};
 
 use byteorder::{BigEndian, WriteBytesExt};
 use ciborium::value::Value;
-use dpp::data_contract::drive_api::DriveContractExt;
+use dpp::data_contract::extra::DriveContractExt;
 use serde::{Deserialize, Serialize};
 
 use crate::common::{bytes_for_system_value_from_tree_map, get_key_from_cbor_map};
-use crate::contract::{Contract, DocumentType};
+use crate::contract::Contract;
 use crate::drive::defaults::PROTOCOL_VERSION;
 use crate::drive::Drive;
+use dpp::data_contract::extra::DocumentType;
+
 use crate::error::contract::ContractError;
 use crate::error::drive::DriveError;
 use crate::error::structure::StructureError;
 use crate::error::Error;
-
-// TODO move the factory to higher higher level abstractions
-// TODO this factory is used in benchmark and tests - so probably the methods should be available under
-// TODO the test feature
-use rand::rngs::StdRng;
-use rand::{Rng, SeedableRng};
-pub trait DocumentFactory {
-    fn random_documents(&self, count: u32, seed: Option<u64>) -> Vec<Document>;
-    fn document_from_bytes(&self, bytes: &[u8]) -> Result<Document, Error>;
-    fn random_document(&self, seed: Option<u64>) -> Document;
-    fn random_document_with_rng(&self, rng: &mut StdRng) -> Document;
-    fn random_filled_documents(&self, count: u32, seed: Option<u64>) -> Vec<Document>;
-    fn random_filled_document(&self, seed: Option<u64>) -> Document;
-    fn random_filled_document_with_rng(&self, rng: &mut StdRng) -> Document;
-}
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
 pub struct Document {
@@ -362,7 +349,8 @@ impl Document {
 mod tests {
     use super::*;
     use crate::common::json_document_to_cbor;
-    use dpp::data_contract::drive_api::DriveContractExt;
+    use crate::contract::DocumentFactory;
+    use dpp::data_contract::extra::DriveContractExt;
 
     #[test]
     fn test_drive_serialization() {
@@ -393,91 +381,6 @@ mod tests {
             let serialized_document = document
                 .serialize_consume(document_type)
                 .expect("expected to serialize");
-        }
-    }
-}
-
-impl DocumentFactory for DocumentType {
-    fn random_documents(&self, count: u32, seed: Option<u64>) -> Vec<Document> {
-        let mut rng = match seed {
-            None => StdRng::from_entropy(),
-            Some(seed_value) => StdRng::seed_from_u64(seed_value),
-        };
-        let mut vec: Vec<Document> = vec![];
-        for _i in 0..count {
-            vec.push(self.random_document_with_rng(&mut rng));
-        }
-        vec
-    }
-
-    fn document_from_bytes(&self, bytes: &[u8]) -> Result<Document, Error> {
-        Document::from_bytes(bytes, self)
-    }
-
-    fn random_document(&self, seed: Option<u64>) -> Document {
-        let mut rng = match seed {
-            None => StdRng::from_entropy(),
-            Some(seed_value) => StdRng::seed_from_u64(seed_value),
-        };
-        self.random_document_with_rng(&mut rng)
-    }
-
-    fn random_document_with_rng(&self, rng: &mut StdRng) -> Document {
-        let id = rng.gen::<[u8; 32]>();
-        let owner_id = rng.gen::<[u8; 32]>();
-        let properties = self
-            .properties
-            .iter()
-            .map(|(key, document_field)| {
-                (key.clone(), document_field.document_type.random_value(rng))
-            })
-            .collect();
-
-        Document {
-            id,
-            properties,
-            owner_id,
-        }
-    }
-
-    fn random_filled_documents(&self, count: u32, seed: Option<u64>) -> Vec<Document> {
-        let mut rng = match seed {
-            None => rand::rngs::StdRng::from_entropy(),
-            Some(seed_value) => rand::rngs::StdRng::seed_from_u64(seed_value),
-        };
-        let mut vec: Vec<Document> = vec![];
-        for _i in 0..count {
-            vec.push(self.random_filled_document_with_rng(&mut rng));
-        }
-        vec
-    }
-
-    fn random_filled_document(&self, seed: Option<u64>) -> Document {
-        let mut rng = match seed {
-            None => rand::rngs::StdRng::from_entropy(),
-            Some(seed_value) => rand::rngs::StdRng::seed_from_u64(seed_value),
-        };
-        self.random_filled_document_with_rng(&mut rng)
-    }
-
-    fn random_filled_document_with_rng(&self, rng: &mut StdRng) -> Document {
-        let id = rng.gen::<[u8; 32]>();
-        let owner_id = rng.gen::<[u8; 32]>();
-        let properties = self
-            .properties
-            .iter()
-            .map(|(key, document_field)| {
-                (
-                    key.clone(),
-                    document_field.document_type.random_filled_value(rng),
-                )
-            })
-            .collect();
-
-        Document {
-            id,
-            properties,
-            owner_id,
         }
     }
 }
