+++ drive/src/drive/document/delete.rs
@@ -10,7 +10,7 @@ use crate::error::drive::DriveError;
 use crate::error::Error;
 use crate::fee::calculate_fee;
 use crate::fee::op::DriveOperation;
-use dpp::data_contract::drive_api::DriveContractExt;
+use dpp::data_contract::extra::DriveContractExt;
 
 impl Drive {
     pub fn delete_document_for_contract(
