+++ drive/src/drive/mod.rs
@@ -24,7 +24,7 @@ pub mod identity;
 pub mod object_size_info;
 pub mod query;
 
-use dpp::data_contract::drive_api::DriveContractExt;
+use dpp::data_contract::extra::DriveContractExt;
 
 pub struct EpochInfo {
     current_epoch: u16,
